// rugchecker/model/result/marketdata-check.ts
export default class MarketdataCheckResult {
  address: string;
  priceSol: number;
  priceUsd: number;
  liquidityUsd: number;
  fdv: number;
  volume24h: number;
  volume6h: number;
  volume1h: number;
  volume5m: number;
  priceChange24h: number;
  priceChange6h: number;
  priceChange1h: number;
  priceChange5m: number;
  buys24h: number;
  buys6h: number;
  buys1h: number;
  buys5m: number;
  sells24h: number;
  sells6h: number;
  sells1h: number;
  sells5m: number;
}